+++ src/Yaat.Sim/Commands/ParsedCommand.cs
@@ -66,7 +66,7 @@ public record ForceAltitudeCommand(int Altitude) : ParsedCommand;
 
 public record ForceSpeedCommand(int Speed) : ParsedCommand;
 
-public record WarpCommand(string PositionLabel, double Latitude, double Longitude, MagneticHeading MagneticHeading, int Altitude, int Speed)
+public record WarpCommand(string PositionLabel, double Latitude, double Longitude, MagneticHeading? MagneticHeading, int? Altitude, int? Speed)
     : ParsedCommand;
 
 public record WarpGroundCommand(string Taxiway1, string Taxiway2, int? NodeId = null, string? ParkingName = null) : ParsedCommand;